requirements.txt
addict==2.4.0
aiohttp==3.6.2
astunparse==1.6.3
async-timeout==3.0.1
attrs==20.2.0
audioread==2.1.9
beautifulsoup4==4.9.2
braceexpand==0.1.5
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
colorama==0.4.3
convertdate==2.3.2
croniter==1.0.12
cycler==0.10.0
discord.py==1.7.3
expiringdict==1.2.1
ffmpeg-normalize==1.20.1
hijri-converter==2.2.2
hiredis==2.0.0
holidays==0.11.3.1
html5lib==1.1
humanize==3.0.0
idna==2.10
import-expression==1.1.4
Jinja2==3.0.1
jishaku==1.19.1.200
kiwisolver==1.3.1
korean-lunar-calendar==0.2.1
MarkupSafe==2.0.1
matplotlib==3.4.2
multidict==4.7.6
numpy==1.20.3
pandas==1.2.4
Pillow==8.4.0
pycparser==2.20
PyMeeus==0.5.11
PyNaCl==1.4.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
redis==3.5.3
redistimeseries==1.4.3
requests==2.24.0
rmtest==0.7.0
rq==1.7.0
rq-scheduler==0.11.0
semantic-version==2.8.5
six==1.15.0
soupsieve==2.0.1
tabulate==0.8.9
toml==0.10.2
tqdm==4.50.0
urbandictionary-top==0.3
urllib3==1.25.10
webencodings==0.5.1
websockets==8.1
yarl==1.6.0
